fix(e2e): address CodeRabbit persist-credentials and parseArgs nits
Disable credential persistence on the CMT initial-status checkout, and
share CLI arg parsing between TSIO status/stub utilities so sparse
checkouts still resolve the dependency.
